Platform

+++ src/windowsX64Main/kotlin/de/takeweiland/konan_test/platform.kt
@@ -8,6 +8,7 @@ actual fun terminalWidth(): Int {
     return tryTerminalWidths(
         { terminalWidthWin32() },
         { terminalWidthTput() }
+
     )
 }
 
@@ -16,7 +17,7 @@ private fun terminalWidthWin32(): Int {
     memScoped {
         val csbi: CONSOLE_SCREEN_BUFFER_INFO = alloc()
         val returnValue = GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), csbi.ptr)
-        return if (returnValue != 0 && (csbi.srWindow.Right != 0 || csbi.srWindow.Left != 0)) {
+        return if (returnValue != 0 && (csbi.srWindow.Right != 0.toShort() || csbi.srWindow.Left != 0.toShort())) {
             csbi.srWindow.Right - csbi.srWindow.Left + 1
         } else {
             -1
